# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/sale-invoices": {
            "get": {
                "tags": [
                    "SaleInvoice"
                ],
                "summary": "Get Sale invoices",
                "operationId": "getSaleInvoices",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SaleInvoice"
                ],
                "summary": "Post all sale-invoices",
                "operationId": "postSaleInvoice",
                "parameters": [
                    {
                        "name": "total_amount",
                        "in": "query",
                        "description": "mobile No",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "discount",
                        "in": "query",
                        "description": "address",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tax",
                        "in": "query",
                        "description": "gender",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "payment_type",
                        "in": "query",
                        "description": "position",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "payment_amount",
                        "in": "query",
                        "description": "shop id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "receive_amount",
                        "in": "query",
                        "description": "shop id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "change",
                        "in": "query",
                        "description": "shop id",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "staff_id",
                        "in": "query",
                        "description": "shop id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "shop id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sale-invoices/{id}": {
            "get": {
                "tags": [
                    "SaleInvoice"
                ],
                "summary": "Show sale-invoices",
                "operationId": "showSaleInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Sale Invoice",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SaleInvoice"
                ],
                "summary": "delete sale-invoices",
                "operationId": "deleteSaleInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id you want to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sale-invoices/get-by-voucher/{voucher_no}": {
            "get": {
                "tags": [
                    "SaleInvoice"
                ],
                "summary": "search by voucher no",
                "operationId": "showSaleInvoiceByVoucher_no",
                "parameters": [
                    {
                        "name": "voucher_no",
                        "in": "path",
                        "description": "voucher no",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}